Orthonormal Search added, multiple optimization were applied

+++ matrix.cpp
@@ -1,8 +1,8 @@
 #include "matrix.h"
 
 Matrix::Matrix(unsigned char* vector,
-               size_t height,
-               size_t width) :
+               int height,
+               int width) :
                _vector(vector),
                _height(height),
                _width(width) {
