+++ matrix.h
@@ -11,15 +11,15 @@ class Matrix {
 public:
     Matrix(
         unsigned char* vector, 
-        size_t height,
-        size_t width
+        int height,
+        int width
     );
     // ISO CPP tells us that if we define function inside the class, eventually
     // compiler makes it inline
-    uint32_t getHeight() const  {
+    int getHeight() const  {
         return this -> _height;
     };
-    uint32_t getWidth() const {
+    int getWidth() const {
         return this -> _width;;
     };
 
@@ -27,8 +27,8 @@ public:
         return static_cast<int>(_vector[h * getWidth() + w]);
     };
 private:
-    uint32_t _height;
-    uint32_t _width;
-    uint32_t _total;
+    int _height;
+    int _width;
+    int _total;
     unsigned char* _vector;
 };
